Update key_generation.md
I was not sure what this meant so couldn't change it:

"But since the hash of a public key is 20 bytes (160 bits), it is smaller than the total universe of the addresses. You might do better."

Perhaps:

"But since the hash of a public key is 20 bytes (160 bits), it is smaller than the total possible number of addresses. You might do better."