Update bitcoin_address.md
I was confused when I was first reading this section by Key privateKey = new Key(); // generate a random private key
new Key() can be said as a private key? I think it's more near to a kind of key object.
So I edit that, if it's not appropriate, I will back to the original one.
And I add some contents for giving background basics about Base64, Base58 etc.